+++ packages/expo-image-manipulator/android/src/main/java/expo/modules/imagemanipulator/ImageRef.kt
@@ -4,4 +4,6 @@ import android.graphics.Bitmap
 import expo.modules.kotlin.RuntimeContext
 import expo.modules.kotlin.sharedobjects.SharedRef
 
-class ImageRef(bitmap: Bitmap, runtimeContext: RuntimeContext) : SharedRef<Bitmap>(bitmap, runtimeContext)
+class ImageRef(bitmap: Bitmap, runtimeContext: RuntimeContext) : SharedRef<Bitmap>(bitmap, runtimeContext) {
+  override val nativeRefType: String = "image"
+}