+++ packages/expo-image/android/src/main/java/expo/modules/image/Image.kt
@@ -3,4 +3,6 @@ package expo.modules.image
 import android.graphics.drawable.Drawable
 import expo.modules.kotlin.sharedobjects.SharedRef
 
-class Image(ref: Drawable) : SharedRef<Drawable>(ref)
+class Image(ref: Drawable) : SharedRef<Drawable>(ref) {
+  override val nativeRefType: String = "image"
+}